+++ packages/adapter-cloudflare/files/worker.js
@@ -10,7 +10,20 @@ export default {
 		const url = new URL(req.url);
 
 		// static assets
-		if (url.pathname.startsWith(prefix)) return env.ASSETS.fetch(req);
+		if (url.pathname.startsWith(prefix)) {
+			/** @type {Response} */
+			const res = await env.ASSETS.fetch(req);
+
+			return new Response(res.body, {
+				headers: {
+					// include original cache headers, minus cache-control which
+					// is overridden, and etag which is no longer useful
+					'cache-control': 'public, immutable, max-age=31536000',
+					'content-type': res.headers.get('content-type'),
+					'x-robots-tag': 'noindex'
+				}
+			});
+		}
 
 		// prerendered pages and index.html files
 		const pathname = url.pathname.replace(/\/$/, '');