+++ packages/kit/test/apps/basics/src/routes/routing/fallthrough-layout/[xyz]/__layout.svelte
@@ -1,0 +1,12 @@
+<script context="module">
+	/** @type {import("@sveltejs/kit").Load} */
+
+	export async function load({ params }) {
+		if (params.xyz !== 'ok') {
+			return { fallthrough: true };
+		}
+		return {};
+	}
+</script>
+
+<slot />