+++ packages/kit/test/apps/basics/test/test.js
@@ -3,9 +3,8 @@ import path from 'path';
 import http from 'http';
 import * as ports from 'port-authority';
 import { expect } from '@playwright/test';
-import { test } from '../../../utils.js';
 import { fileURLToPath } from 'url';
-import { start_server } from '../../../utils.js';
+import { start_server, test } from '../../../utils.js';
 
 /** @typedef {import('@playwright/test').Response} Response */
 
@@ -89,10 +88,8 @@ test.describe.parallel('a11y', () => {
 });
 
 test.describe('Scrolling', () => {
-	// skip these tests if
-	// a) JS is disabled, since we're testing client-side behaviour, or
-	// b) we're in CI, because for unknown reasons the tests are flaky as hell there
-	test.skip(({ javaScriptEnabled }) => !javaScriptEnabled || !!process.env.CI);
+	// skip these tests if JS is disabled, since we're testing client-side behaviour
+	test.skip(({ javaScriptEnabled }) => !javaScriptEnabled);
 
 	test('url-supplied anchor works on direct page load', async ({ page, in_view }) => {
 		await page.goto('/anchor/anchor#go-to-element');
@@ -145,11 +142,9 @@ test.describe('Scrolling', () => {
 
 	test('url-supplied anchor is ignored with onMount() scrolling on direct page load', async ({
 		page,
-		in_view,
-		started
+		in_view
 	}) => {
 		await page.goto('/anchor-with-manual-scroll/anchor#go-to-element');
-		await started();
 		expect(await in_view('#abcde')).toBe(true);
 	});
 
@@ -620,6 +615,16 @@ test.describe.parallel('Errors', () => {
 			expect(await page.innerHTML('h1')).toBe('401');
 		}
 	});
+
+	test('error thrown in handle results in a rendered error page', async ({ page }) => {
+		await page.goto('/errors/error-in-handle');
+
+		expect(await page.textContent('footer')).toBe('Custom layout');
+		expect(await page.textContent('#message')).toBe(
+			'This is your custom error page saying: "Error in handle"'
+		);
+		expect(await page.innerHTML('h1')).toBe('500');
+	});
 });
 
 test.describe.parallel('ETags', () => {
@@ -1077,6 +1082,17 @@ test.describe.parallel('$app/paths', () => {
 			})
 		);
 	});
+
+	test('replaces %svelte.assets% in template with relative path', async ({ page }) => {
+		await page.goto('/');
+		expect(await page.getAttribute('link[rel=icon]', 'href')).toBe('./favicon.png');
+
+		await page.goto('/routing');
+		expect(await page.getAttribute('link[rel=icon]', 'href')).toBe('./favicon.png');
+
+		await page.goto('/routing/rest/foo/bar/baz');
+		expect(await page.getAttribute('link[rel=icon]', 'href')).toBe('../../../../favicon.png');
+	});
 });
 
 test.describe.parallel('$app/stores', () => {
@@ -1097,6 +1113,29 @@ test.describe.parallel('$app/stores', () => {
 
 		const oops = await page.evaluate(() => window.oops);
 		expect(oops).toBeUndefined();
+	});
+
+	test('page store contains stuff', async ({ page, clicknav }) => {
+		await page.goto('/store/stuff/www');
+
+		expect(await page.textContent('#store-stuff')).toBe(
+			JSON.stringify({ name: 'SvelteKit', value: 456, page: 'www' })
+		);
+
+		await clicknav('a[href="/store/stuff/zzz"]');
+		expect(await page.textContent('#store-stuff')).toBe(
+			JSON.stringify({ name: 'SvelteKit', value: 456, page: 'zzz' })
+		);
+
+		await clicknav('a[href="/store/stuff/xxx"]');
+		expect(await page.textContent('#store-stuff')).toBe(
+			JSON.stringify({ name: 'SvelteKit', value: 789, error: 'Params = xxx' })
+		);
+
+		await clicknav('a[href="/store/stuff/yyy"]');
+		expect(await page.textContent('#store-stuff')).toBe(
+			JSON.stringify({ name: 'SvelteKit', value: 789, error: 'Params = yyy' })
+		);
 	});
 
 	test('navigating store contains from and to', async ({ app, page, javaScriptEnabled }) => {
@@ -1219,11 +1258,10 @@ test.describe.parallel('Redirects', () => {
 		);
 	});
 
-	test('redirect-on-load', async ({ baseURL, page, javaScriptEnabled, started }) => {
+	test('redirect-on-load', async ({ baseURL, page, javaScriptEnabled }) => {
 		await page.goto('/redirect-on-load');
 
 		if (javaScriptEnabled) {
-			await started();
 			await page.waitForTimeout(50); // TODO investigate why this test is flaky
 			expect(page.url()).toBe(`${baseURL}/redirect-on-load/redirected`);
 			expect(await page.textContent('h1')).toBe('Hazaa!');
@@ -1239,8 +1277,7 @@ test.describe.parallel('Routing', () => {
 		page,
 		clicknav,
 		app,
-		javaScriptEnabled,
-		started
+		javaScriptEnabled
 	}) => {
 		await page.goto('/routing/slashes');
 
@@ -1250,7 +1287,6 @@ test.describe.parallel('Routing', () => {
 
 		if (javaScriptEnabled) {
 			await page.goto(`${baseURL}/routing/slashes`);
-			await started();
 			await app.goto('/routing/');
 			expect(page.url()).toBe(`${baseURL}/routing`);
 			expect(await page.textContent('h1')).toBe('Great success!');
@@ -1262,8 +1298,7 @@ test.describe.parallel('Routing', () => {
 		page,
 		clicknav,
 		app,
-		javaScriptEnabled,
-		started
+		javaScriptEnabled
 	}) => {
 		await page.goto('/routing/slashes');
 
@@ -1273,7 +1308,6 @@ test.describe.parallel('Routing', () => {
 
 		if (javaScriptEnabled) {
 			await page.goto(`${baseURL}/routing/slashes`);
-			await started();
 			await app.goto('/routing/?');
 			expect(page.url()).toBe(`${baseURL}/routing`);
 			expect(await page.textContent('h1')).toBe('Great success!');
@@ -1285,8 +1319,7 @@ test.describe.parallel('Routing', () => {
 		page,
 		clicknav,
 		app,
-		javaScriptEnabled,
-		started
+		javaScriptEnabled
 	}) => {
 		await page.goto('/routing/slashes');
 
@@ -1296,7 +1329,6 @@ test.describe.parallel('Routing', () => {
 
 		if (javaScriptEnabled) {
 			await page.goto(`${baseURL}/routing/slashes`);
-			await started();
 			await app.goto('/routing/?foo=bar');
 			expect(page.url()).toBe(`${baseURL}/routing?foo=bar`);
 			expect(await page.textContent('h1')).toBe('Great success!');
@@ -1334,13 +1366,11 @@ test.describe.parallel('Routing', () => {
 		app,
 		page,
 		clicknav,
-		javaScriptEnabled,
-		started
+		javaScriptEnabled
 	}) => {
 		if (javaScriptEnabled) {
 			await page.goto('/routing');
 
-			await started();
 			await app.prefetchRoutes(['/routing/a']).catch((e) => {
 				// from error handler tests; ignore
 				if (!e.message.includes('Crashing now')) throw e;
@@ -1357,25 +1387,17 @@ test.describe.parallel('Routing', () => {
 		}
 	});
 
-	test('navigates programmatically', async ({ page, app, javaScriptEnabled, started }) => {
+	test('navigates programmatically', async ({ page, app, javaScriptEnabled }) => {
 		if (javaScriptEnabled) {
 			await page.goto('/routing/a');
-			await started();
 			await app.goto('/routing/b');
 			expect(await page.textContent('h1')).toBe('b');
 		}
 	});
 
-	test('prefetches programmatically', async ({
-		baseURL,
-		page,
-		app,
-		javaScriptEnabled,
-		started
-	}) => {
+	test('prefetches programmatically', async ({ baseURL, page, app, javaScriptEnabled }) => {
 		if (javaScriptEnabled) {
 			await page.goto('/routing/a');
-			await started();
 
 			/** @type {string[]} */
 			let requests = [];
@@ -1500,6 +1522,25 @@ test.describe.parallel('Routing', () => {
 		).toBe('rgb(255, 0, 0)');
 	});
 
+	test('$page.url.hash is correctly set on page load', async ({ page, javaScriptEnabled }) => {
+		if (javaScriptEnabled) {
+			await page.goto('/routing/hashes/pagestore#target');
+			expect(await page.textContent('#window-hash')).toBe('#target');
+			expect(await page.textContent('#page-url-hash')).toBe('#target');
+		}
+	});
+
+	test('$page.url.hash is correctly set on navigation', async ({ page, javaScriptEnabled }) => {
+		if (javaScriptEnabled) {
+			await page.goto('/routing/hashes/pagestore');
+			expect(await page.textContent('#window-hash')).toBe('');
+			expect(await page.textContent('#page-url-hash')).toBe('');
+			await page.click('[href="#target"]');
+			expect(await page.textContent('#window-hash')).toBe('#target');
+			expect(await page.textContent('#page-url-hash')).toBe('#target');
+		}
+	});
+
 	test('fallthrough', async ({ page }) => {
 		await page.goto('/routing/fallthrough-simple/invalid');
 		expect(await page.textContent('h1')).toBe('Page');
@@ -1513,6 +1554,17 @@ test.describe.parallel('Routing', () => {
 		expect(await page.textContent('h1')).toBe('camel is an animal');
 
 		await clicknav('[href="/routing/fallthrough-advanced/potato"]');
+		expect(await page.textContent('h1')).toBe('404');
+	});
+
+	test('dynamic fallthrough of layout', async ({ page, clicknav }) => {
+		await page.goto('/routing/fallthrough-layout/okay');
+		expect(await page.textContent('h1')).toBe('foo is okay');
+
+		await clicknav('[href="/routing/fallthrough-layout/ok"]');
+		expect(await page.textContent('h1')).toBe('xyz is ok');
+
+		await clicknav('[href="/routing/fallthrough-layout/notok"]');
 		expect(await page.textContent('h1')).toBe('404');
 	});
 
@@ -1619,6 +1671,11 @@ test.describe.parallel('Routing', () => {
 		await clicknav('[href="/routing/rest/path/three"]');
 		expect(await page.textContent('h1')).toBe('path: /routing/rest/path/three');
 	});
+
+	test('allows rest routes to have prefixes and suffixes', async ({ page }) => {
+		await page.goto('/routing/rest/complex/prefix-one/two/three');
+		expect(await page.textContent('h1')).toBe('parts: one/two/three');
+	});
 });
 
 test.describe.parallel('Session', () => {
@@ -1641,13 +1698,11 @@ test.describe.parallel('Shadow DOM', () => {
 		app,
 		page,
 		clicknav,
-		javaScriptEnabled,
-		started
+		javaScriptEnabled
 	}) => {
 		await page.goto('/routing/shadow-dom');
 
 		if (javaScriptEnabled) {
-			await started();
 			await app.prefetchRoutes(['/routing/a']).catch((e) => {
 				// from error handler tests; ignore
 				if (!e.message.includes('Crashing now')) throw e;