+++ packages/adapter-cloudflare/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sveltejs/adapter-cloudflare
 
+## 1.0.0-next.7
+
+### Patch Changes
+
+- Add immutable cache headers to generated assets ([#3222](https://github.com/sveltejs/kit/pull/3222))
+
+## 1.0.0-next.6
+
+### Patch Changes
+
+- use posix to resolve relative path ([#3214](https://github.com/sveltejs/kit/pull/3214))
+
 ## 1.0.0-next.5
 
 ### Patch Changes